break the renamio function out and have a custom windows version
